requirements.txt
openai
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-ollama
langchain-huggingface
langsmith
huggingface-hub
transformers
torch
python-dotenv
langgraph
bitsandbytes
accelerate
tiktoken
beautifulsoup4